+++ src/test/groovy/uk/q3c/krail/core/guice/DefaultBindingManager_IntegrationTest.groovy
@@ -15,7 +15,6 @@ package uk.q3c.krail.core.guice
 
 import com.google.inject.Module
 import com.google.inject.TypeLiteral
-import uk.q3c.krail.core.option.OptionContainerProvider
 import uk.q3c.krail.core.option.inmemory.InMemoryOptionContainerProvider
 import uk.q3c.krail.core.validation.JavaxValidationSubstitutes
 import uk.q3c.krail.core.validation.KrailInterpolator
@@ -23,6 +22,7 @@ import uk.q3c.krail.i18n.I18NKey
 import uk.q3c.krail.i18n.persist.I18NPersistenceHelper
 import uk.q3c.krail.option.inmemory.dao.InMemoryOptionDaoDelegate
 import uk.q3c.krail.option.persist.ActiveOptionSourceDefault
+import uk.q3c.krail.option.persist.OptionContainerProvider
 import uk.q3c.krail.option.persist.OptionDaoDelegate
 import uk.q3c.krail.persist.inmemory.InMemory
 import uk.q3c.krail.testutil.dummy.Dummy